pyproject.toml
[tool.poetry]
name = "api-gateway-users"
version = "0.1.0"
description = ""
authors = ["Vinicius de Andrade <viniciusdeandrade@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
django = "5.1.*"
django-unfold = "0.46.*"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
